Moving pageheader to ItemsPage
