+++ WinUIGallery/AllControlsPage.xaml
@@ -32,7 +32,6 @@
                         Until the XamlBehaviors are updated to work with lifted XAML, just use the smallHeaderText all the time.
                     -->
                     <VisualState.Setters>
-                        <Setter Target="pageHeader.Visibility" Value="Collapsed" />
                         <Setter Target="smallHeaderText.Visibility" Value="Visible"/>
                     </VisualState.Setters>
                 </VisualState>
@@ -45,7 +44,6 @@
                         <Setter Target="itemGridView.Margin" Value="0,-44,0,0" />
                         <Setter Target="headerImageView.Margin" Value="-12,0,-12,12" />
                         <Setter Target="itemGridView.ItemContainerStyle" Value="{StaticResource GridViewItemStyleSmall}" />
-                        <Setter Target="pageHeader.Visibility" Value="Collapsed" />
                         <Setter Target="smallHeaderText.Visibility" Value="Visible"/>
                     </VisualState.Setters>
                 </VisualState>
@@ -67,11 +65,6 @@
             Loaded="OnItemGridViewLoaded"
             SizeChanged="OnItemGridViewSizeChanged"
             SelectionMode="None">
-            <!-- XAML_BEVHAVIORS_REMOVED
-            <interactivity:Interaction.Behaviors>
-                <behavior:ImageScrollBehavior TargetControl="{Binding ElementName=pageHeader}" />
-            </interactivity:Interaction.Behaviors>
-            -->
 
             <GridView.Header>
                 <Grid x:Name="HeaderGrid" Margin="12,39,43,30">
@@ -98,16 +91,5 @@
                 </Grid>
             </GridView.Header>
         </GridView>
-
-        <local:PageHeader x:Name="pageHeader"
-            Title="{StaticResource ControlsName}"
-            VerticalAlignment="Top"
-            Margin="44,92"
-            Foreground="black"
-            Background="{ThemeResource HomePageBackgroundBrush}"
-            FontSize="28"
-            Padding="{ThemeResource NavigationViewHeaderMargin}"
-            IsHitTestVisible="False"
-            ShadowOpacity="0"/>
     </Grid>
 </local:ItemsPageBase>
