+++ WinUIGallery/ControlPages/AppBarSeparatorPage.xaml.cs
@@ -1,4 +1,4 @@
-﻿//*********************************************************
+//*********************************************************
 //
 // Copyright (c) Microsoft. All rights reserved.
 // THIS CODE IS PROVIDED *AS IS* WITHOUT WARRANTY OF
@@ -28,10 +28,11 @@ namespace AppUIBasics.ControlPages
 
         private void AppBarSeparatorPage_Unloaded(object sender, RoutedEventArgs e)
         {
-            CommandBar appBar = NavigationRootPage.GetForElement(this).PageHeader.TopCommandBar;
-            compactButton.Click -= CompactButton_Click;
-            appBar.PrimaryCommands.Remove(compactButton);
-            appBar.PrimaryCommands.Remove(separator);
+         
+            //CommandBar appBar = NavigationRootPage.GetForElement(this).PageHeader.TopCommandBar;
+            //compactButton.Click -= CompactButton_Click;
+            //appBar.PrimaryCommands.Remove(compactButton);
+            //appBar.PrimaryCommands.Remove(separator);
         }
 
         void AppBarButtonPage_Loaded(object sender, RoutedEventArgs e)
@@ -39,17 +40,17 @@ namespace AppUIBasics.ControlPages
             // Add compact button to the command bar. It provides functionality specific
             // to this page, and is removed when leaving the page.
 
-            CommandBar appBar = NavigationRootPage.GetForElement(this).PageHeader.TopCommandBar;
-            separator = new AppBarSeparator();
-            appBar.PrimaryCommands.Insert(0, separator);
+            //CommandBar appBar = NavigationRootPage.GetForElement(this).PageHeader.TopCommandBar;
+            //separator = new AppBarSeparator();
+            //appBar.PrimaryCommands.Insert(0, separator);
 
-            compactButton = new AppBarToggleButton
-            {
-                Icon = new SymbolIcon(Symbol.FontSize),
-                Label = "IsCompact"
-            };
-            compactButton.Click += CompactButton_Click;
-            appBar.PrimaryCommands.Insert(0, compactButton);
+            //compactButton = new AppBarToggleButton
+            //{
+            //    Icon = new SymbolIcon(Symbol.FontSize),
+            //    Label = "IsCompact"
+            //};
+            //compactButton.Click += CompactButton_Click;
+            //appBar.PrimaryCommands.Insert(0, compactButton);
         }
 
         private void CompactButton_Click(object sender, RoutedEventArgs e)
