--- WinUIGallery/PageHeader.xaml
@@ -1,128 +1,0 @@
-<UserControl
-    x:Class="AppUIBasics.PageHeader"
-    xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
-    xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
-    xmlns:local="using:AppUIBasics"
-    x:Name="headerControl"
-    Background="Transparent"
-    Foreground="{ThemeResource SystemControlForegroundBaseHighBrush}"
-    FontSize="24"
-    HighContrastAdjustment="Auto">
-    <Grid>
-        <VisualStateManager.VisualStateGroups>
-            <VisualStateGroup x:Name="TitleForegroundStates">
-                <VisualState x:Name="DefaultForeground" />
-                <VisualState x:Name="AlternateForeground">
-                    <VisualState.Setters>
-                        <Setter Target="TitleTextBlock.Foreground" Value="Transparent" />
-                        <Setter Target="ShadowHost.Opacity" Value="0"/>
-                    </VisualState.Setters>
-                </VisualState>
-            </VisualStateGroup>
-        </VisualStateManager.VisualStateGroups>
-
-        <Rectangle x:Name="NarrowBackground" Visibility="Collapsed" />
-
-        <Rectangle x:Name="WideBackground" Fill="{Binding ElementName=headerControl, Path=Background}" Opacity="{Binding ElementName=headerControl, Path=BackgroundColorOpacity}" />
-
-        <Grid x:Name="headerRoot" Padding="{Binding ElementName=headerControl, Path=Padding}" VerticalAlignment="Top">
-            <Grid.ColumnDefinitions>
-                <ColumnDefinition Width="*" />
-                <ColumnDefinition Width="Auto" />
-            </Grid.ColumnDefinitions>
-            <Grid x:Name="pageTitle"
-                Background="Transparent"
-                RowDefinitions="Auto,Auto"
-                VerticalAlignment="Top">
-                <Canvas x:Name="ShadowHost" Grid.RowSpan="2" Opacity="{x:Bind ShadowOpacity, Mode=OneWay}"/>
-                <TextBlock x:Name="TitleTextBlock"
-                    Style="{StaticResource TitleTextBlockStyle}"
-                    VerticalAlignment="Center"
-                    FontSize="{Binding ElementName=headerControl, Path=FontSize}"
-                    Foreground="{Binding ElementName=headerControl, Path=Foreground}"
-                    Text="{x:Bind Title, Mode=OneWay}"
-                    TextTrimming="CharacterEllipsis"
-                    TextWrapping="NoWrap" />
-                <TextBlock x:Name="NamespaceTextBlock" Grid.Row="1" Text="{x:Bind ApiNamespace, Mode=OneWay}"
-                    Visibility="{x:Bind ApiNamespace, Converter={StaticResource emptyStringToVisibilityConverter}, Mode=OneWay}"
-                    Style="{StaticResource SubtitleTextBlockStyle}"
-                    Margin="0,0,0,6"
-                    FontSize="14" FontStyle="Italic"
-                    Foreground="{ThemeResource SystemControlForegroundBaseMediumBrush}"/>
-            </Grid>
-            <Border x:Name="commandBarBorder" Grid.Column="1" VerticalAlignment="Top">
-                <CommandBar
-                    x:Name="topCommandBar"
-                    Height="40"
-                    Margin="0,0,12,0"
-                    Background="Transparent"
-                    ClosedDisplayMode="Compact"
-                    DefaultLabelPosition="Right"
-                    IsOpen="False"
-                    IsSticky="False"
-                    Visibility="Collapsed">
-                    <AppBarButton x:Name="CopyLinkButton" Click="OnCopyLinkButtonClick">
-                        <AppBarButton.Icon>
-                            <SymbolIcon x:Name="CopyLinkButtonIcon" Margin="0,-2,0,0" Symbol="Link" />
-                        </AppBarButton.Icon>
-                        <AppBarButton.Resources>
-                            <TeachingTip
-                                x:Name="CopyLinkButtonTeachingTip"
-                                Title="Quickly reference this sample!"
-                                PreferredPlacement="Bottom"
-                                Subtitle="Share with others or paste this link into the Run dialog to open the app to this page directly."
-                                ActionButtonContent="Don't show again"
-                                ActionButtonClick="OnCopyDontShowAgainButtonClick"
-                                CloseButtonContent="Got it!"
-                                Target="{x:Bind CopyLinkButton}">
-                                <TeachingTip.HeroContent>
-                                    <Image Source="/Assets/CopyLinkTeachingTip.png" />
-                                </TeachingTip.HeroContent>
-                            </TeachingTip>
-                        </AppBarButton.Resources>
-                    </AppBarButton>
-                    <AppBarButton
-                        x:Name="ThemeButton"
-                        Click="OnThemeButtonClick"
-                        Label="Toggle Theme">
-                        <AppBarButton.Icon>
-                            <FontIcon Glyph="&#xE771;" Margin="0,-2,0,0" />
-                        </AppBarButton.Icon>
-                        <AppBarButton.Resources>
-                            <TeachingTip x:Name="ToggleThemeTeachingTip1"                   
-                                Target="{x:Bind ThemeButton}"
-                                Title="Change themes without hassle"                                        
-                                Subtitle="It's easier than ever to see control samples in both light and dark theme!">
-                                <TeachingTip.IconSource>
-                                    <SymbolIconSource Symbol="Refresh" />
-                                </TeachingTip.IconSource>
-                            </TeachingTip>
-                            <TeachingTip x:Name="ToggleThemeTeachingTip2"                   
-                                Title="Change themes without hassle"
-                                Subtitle="It's easier than ever to see control samples in both light and dark theme!"
-                                PreferredPlacement="Auto"                                              
-                                PlacementMargin="20"
-                                IsLightDismissEnabled="True"
-                                ActionButtonClick="ToggleThemeTeachingTip2_ActionButtonClick"
-                                ActionButtonContent="Toggle theme now"
-                                CloseButtonContent="Got it!">
-                            </TeachingTip>
-                            <TeachingTip x:Name="ToggleThemeTeachingTip3" 
-                                Target="{x:Bind ThemeButton}"
-                                Title="Change themes without hassle"
-                                Subtitle="It's easier than ever to see control samples in both light and dark theme!"
-                                PreferredPlacement="LeftBottom">
-                                <TeachingTip.HeroContent>
-                                    <Image Source="/Assets/SampleMedia/sunset.jpg" AutomationProperties.Name="Sunset" />
-                                </TeachingTip.HeroContent>
-                                <TeachingTip.Content>
-                                    <TextBlock TextWrapping="WrapWholeWords" Margin="0,16,0,0">To change your desktop theme visit the <Hyperlink NavigateUri="ms-settings:colors">Settings app</Hyperlink>.</TextBlock>
-                                </TeachingTip.Content>
-                            </TeachingTip>
-                        </AppBarButton.Resources>
-                    </AppBarButton>
-                </CommandBar>
-            </Border>
-        </Grid>
-    </Grid>
-</UserControl>
